Change Contribution Limit Logic
Rather than allow for contributors to exceed the maxGoal, we consider the maxGoal a hard limit.

This creates a new scenario where the minContribution is greater than the delta between max goal and
total raised funds. However, we accept this constraint and note that existing contributors can top it
off to expedite processing.

Reworked documentation as well.