[STP] Fix slashing bug
Slashing rewards of an account which previously redeemed rewards resulted in redemption values greater than
the reward pool balance.

This change decrements the withdrawn reward state variable for a given account, as well as increases a slash pool
variable to make a portion of the reward pool ineligible for redemption.