[STP] Reward Pool Fixes
In the event that the creator deploys a contract with reward bps and the first mint occurs after the reward multiplier reaches 0, funds would be allocated to the reward pool, yet no one could withdraw.

Additionally, given slashing mechanics, it was possible that total points reached 0, leaving a reward pool balance greater than 0, with no way to recover.

This commit address both of those scenarios.