[STP] QS Redesign slashing mechanism
Slashing points proportially to inactive time resulted in truncation issues between slashed points and withdraws due
to differing resolution. While the impact may have been minimal, it was imperfect and increased uncertainty.

This model is simple: Points can be slashed once for a given account, after a grace period which is 50% of their total purchased
time. If an account subscribes, in aggregate, for 6 months, then their grace period is 3 months. After the grace period, their reward
points can be slashed. If they renew before the slash point, the slash point is reset.

Additionally, the slashing point will factor in grants, but only if grant time is active.